refactor: LPセクションを components へ移動しデモをリネーム
Copilot レビュー指摘に従い、components→app の逆依存を解消し
HomeCompactDemo を MaskingCompactDemo に改名する。
